Emit votes_changed signal after reading votes and fix button gets enabled direct after vote
The problem with the button was, that after a vote the balance_status changed signal got emitted and the table updates. We "fixed" this problem by checking if the balance_status really changed, but if the table resets because of another event we will have this problem again. A good solution would be to check for every button if the user already voted, but the vote was not already mined.
Fixes #151